refactor(report): Tách component
