+++ src/pages/HR/Reports.jsx
@@ -1,16 +1,18 @@
 import React, { useContext, useEffect, useMemo, useState } from "react";
 import HRSidebar from "../../components/Layout/HRSidebar";
-import Modal from "../../components/Layout/Modal";
 import { AuthContext } from "../../context/AuthContext";
 import hrApi from "../../api/hrApi";
 import reportApi from "../../api/reportApi";
 import { exportFinalReportsToExcel } from "../../utils/excelExport";
 import { toast } from "react-toastify";
-import Pagination from "../../components/Common/Pagination";
 import "../../styles/manageInterns.css";
 import "../../styles/table.css";
 import "../../styles/allowances.css";
 import "../../styles/buttons.css";
+import ReportStats from "./Reports/component/ReportStats";
+import ReportFilters from "./Reports/component/ReportFilters";
+import ReportTable from "./Reports/component/ReportTable";
+import InternDetailModal from "./Reports/modals/InternDetailModal";
 
 const ITEMS_PER_PAGE = 10;
 
@@ -464,435 +466,52 @@ const HRReports = () => {
       <div className="dashboard-content">
         <h2 className="page-title">Báo cáo đánh giá cuối kỳ</h2>
 
-        {report && (
-          <div className="stats-row">
-            <div className="stat-card">
-              <div>
-                <h4>Tổng số thực tập sinh</h4>
-                <p className="stat-value">{report.totalInterns ?? 0}</p>
-              </div>
-            </div>
-            <div className="stat-card">
-              <div>
-                <h4>Số thực tập sinh đã được đánh giá</h4>
-                <p className="stat-value">{report.internsWithEvaluations ?? 0}</p>
-              </div>
-            </div>
-            <div className="stat-card">
-              <div>
-                <h4>Điểm trung bình cuối kỳ</h4>
-                <p className="stat-value">
-                  {report.avgFinalScore != null
-                    ? report.avgFinalScore.toFixed(2)
-                    : "-"}
-                </p>
-              </div>
-            </div>
-          </div>
-        )}
+        <ReportStats report={report} />
 
-        <div className="filter-container">
-          <div className="filter-row">
-            <div className="filter-group">
-              <label>Tìm kiếm (tên / SĐT / email)</label>
-              <input
-                type="text"
-                value={pendingSearchKeyword}
-                onChange={(e) => setPendingSearchKeyword(e.target.value)}
-                placeholder="Nhập tên, SĐT hoặc email"
-              />
-            </div>
-            <div className="filter-group">
-              <label>Chương trình</label>
-              <select
-                value={pendingProgramId}
-                onChange={(e) => {
-                  const value = e.target.value;
-                  setPendingProgramId(value);
-                  setSelectedTeamId("");
-                  setPendingTeamName("");
-                  setTeams([]);
-                }}
-              >
-                <option value="">-- Tất cả chương trình --</option>
-                {programs.map((p) => (
-                  <option key={p.programId} value={p.programId}>
-                    {p.name}
-                  </option>
-                ))}
-              </select>
-            </div>
+        <ReportFilters
+          programs={programs}
+          pendingSearchKeyword={pendingSearchKeyword}
+          setPendingSearchKeyword={setPendingSearchKeyword}
+          pendingProgramId={pendingProgramId}
+          setPendingProgramId={setPendingProgramId}
+          pendingMajor={pendingMajor}
+          setPendingMajor={setPendingMajor}
+          pendingTeamName={pendingTeamName}
+          setPendingTeamName={setPendingTeamName}
+          pendingMentorName={pendingMentorName}
+          setPendingMentorName={setPendingMentorName}
+          majorOptions={majorOptions}
+          mentorOptions={mentorOptions}
+          teams={teams}
+          sortedTeams={sortedTeams}
+          report={report}
+          loading={loading}
+          token={token}
+          exporting={exporting}
+          handleSearchReport={handleSearchReport}
+          handleExport={handleExport}
+          handleClearFilters={handleClearFilters}
+          setSelectedTeamId={setSelectedTeamId}
+          setTeams={setTeams}
+        />
 
-            <div className="filter-group">
-              <label>Ngành (tuỳ chọn)</label>
-              <select
-                value={pendingMajor}
-                onChange={(e) => setPendingMajor(e.target.value)}
-                disabled={!report?.interns?.length}
-              >
-                <option value="">-- Tất cả ngành --</option>
-                {majorOptions.map((major) => (
-                  <option key={major} value={major}>
-                    {major}
-                  </option>
-                ))}
-              </select>
-            </div>
-
-            <div className="filter-group">
-              <label>Team (tuỳ chọn)</label>
-              <select
-                value={pendingTeamName}
-                onChange={(e) => {
-                  const name = e.target.value;
-                  setPendingTeamName(name);
-                  const team = sortedTeams.find((t, idx) => {
-                    const generatedName = `Team ${idx + 1}`;
-                    return generatedName === name;
-                  });
-                  setSelectedTeamId(team ? team.teamId : "");
-                }}
-                disabled={!teams.length}
-              >
-                <option value="">-- Tất cả team --</option>
-                {sortedTeams.map((t, idx) => {
-                  const name = `Team ${idx + 1}`;
-                  return (
-                    <option key={t.teamId} value={name}>
-                      {name}
-                    </option>
-                  );
-                })}
-              </select>
-            </div>
-
-            <div className="filter-group">
-              <label>Mentor (tuỳ chọn)</label>
-              <select
-                value={pendingMentorName}
-                onChange={(e) => setPendingMentorName(e.target.value)}
-                disabled={!report?.interns?.length}
-              >
-                <option value="">-- Tất cả mentor --</option>
-                {mentorOptions.map((name) => (
-                  <option key={name} value={name}>
-                    {name}
-                  </option>
-                ))}
-              </select>
-            </div>
-
-            <div className="filter-actions">
-              <button
-                className="btn-filter-apply"
-                onClick={() => handleSearchReport(false)}
-                disabled={loading || !token}
-              >
-                {loading ? "Đang tải..." : "Tìm kiếm"}
-              </button>
-              <button
-                className="btn-filter-apply"
-                onClick={handleExport}
-                disabled={exporting || !report}
-              >
-                {exporting ? "Đang xuất..." : "Xuất Excel"}
-              </button>
-            </div>
-          </div>
-          <div className="clear-filter-container">
-            <button className="clear-filter-btn" onClick={handleClearFilters}>
-              ✖ Clear filter
-            </button>
-          </div>
-        </div>
-
-        {report && (
-          <>
-            <div className="users-table-container">
-              <table className="users-table">
-                <thead>
-                  <tr>
-                    <th>STT</th>
-                    <th>Họ tên</th>
-                    <th>Email</th>
-                    <th>SĐT</th>
-                    <th>Trường</th>
-                    <th>Ngành</th>
-                    {!appliedProgramId && <th>Chương trình</th>}
-                    <th>Nhóm</th>
-                    <th>Mentor</th>
-                    <th>Kỹ thuật</th>
-                    <th>Giao tiếp</th>
-                    <th>Kỷ luật</th>
-                    <th>Thái độ</th>
-                    <th>Điểm cuối kỳ</th>
-                    <th>Ghi chú</th>
-                  </tr>
-                </thead>
-                <tbody>
-                  {paginatedInternRows.length === 0 && (
-                    <tr>
-                      <td colSpan="15" style={{ textAlign: "center" }}>
-                        Không có dữ liệu báo cáo
-                      </td>
-                    </tr>
-                  )}
-                  {paginatedInternRows.map((intern, index) => (
-                    <tr key={intern.internId}>
-                      <td>{(currentPage - 1) * ITEMS_PER_PAGE + index + 1}</td>
-                      <td>{intern.fullName}</td>
-                      <td>{intern.email}</td>
-                      <td>{intern.phone}</td>
-                      <td>{intern.school}</td>
-                      <td>{intern.major}</td>
-                      {!appliedProgramId && (
-                        <td>{intern.programName}</td>
-                      )}
-                      <td>{intern.teamName}</td>
-                      <td>{intern.mentorName}</td>
-                      <td>
-                        {intern.avgTechnical != null
-                          ? intern.avgTechnical.toFixed(2)
-                          : "-"}
-                      </td>
-                      <td>
-                        {intern.avgCommunication != null
-                          ? intern.avgCommunication.toFixed(2)
-                          : "-"}
-                      </td>
-                      <td>
-                        {intern.avgDiscipline != null
-                          ? intern.avgDiscipline.toFixed(2)
-                          : "-"}
-                      </td>
-                      <td>
-                        {intern.avgAttitude != null
-                          ? intern.avgAttitude.toFixed(2)
-                          : "-"}
-                      </td>
-                      <td>
-                        {intern.finalScore != null
-                          ? intern.finalScore.toFixed(2)
-                          : "-"}
-                      </td>
-                      <td
-                        title={intern.allNotes || ""}
-                        style={{
-                          cursor:
-                            intern.allNotes || intern.latestNote
-                              ? "pointer"
-                              : "default",
-                          color:
-                            intern.allNotes || intern.latestNote
-                              ? "#01579b"
-                              : "inherit",
-                          textDecoration:
-                            intern.allNotes || intern.latestNote
-                              ? "underline"
-                              : "none",
-                        }}
-                        onClick={() => {
-                          if (intern.allNotes || intern.latestNote) {
-                            setSelectedIntern(intern);
-                          }
-                        }}
-                      >
-                        {intern.latestNote ||
-                          (intern.allNotes ? "Xem chi tiết" : "-")}
-                      </td>
-                    </tr>
-                  ))}
-                </tbody>
-              </table>
-            </div>
-            <Pagination
-              currentPage={currentPage}
-              totalPages={totalPages}
-              totalItems={internRows.length}
-              onPageChange={setCurrentPage}
-            />
-          </>
-        )}
+        <ReportTable
+          report={report}
+          appliedProgramId={appliedProgramId}
+          paginatedInternRows={paginatedInternRows}
+          currentPage={currentPage}
+          internRows={internRows}
+          totalPages={totalPages}
+          setCurrentPage={setCurrentPage}
+          itemsPerPage={ITEMS_PER_PAGE}
+          setSelectedIntern={setSelectedIntern}
+        />
       </div>
-      {selectedIntern && (
-        <Modal
-          title="Chi tiết đánh giá"
-          onClose={() => setSelectedIntern(null)}
-        >
-          <div style={{ display: "flex", flexDirection: "column", gap: 20 }}>
-            <div
-              style={{
-                borderBottom: "1px solid #e0e0e0",
-                paddingBottom: 10,
-                marginBottom: 4,
-              }}
-            >
-              <h4
-                style={{
-                  margin: 0,
-                  fontSize: 16,
-                  fontWeight: 600,
-                  color: "#374151",
-                }}
-              >
-                Thông tin thực tập sinh
-              </h4>
-            </div>
 
-            <div
-              style={{
-                display: "grid",
-                gridTemplateColumns: "repeat(2, minmax(0, 1fr))",
-                gap: 8,
-                rowGap: 4,
-              }}
-            >
-              <p style={{ margin: 0 }}>
-                <strong>Họ tên:</strong> {selectedIntern.fullName || "-"}
-              </p>
-              <p style={{ margin: 0 }}>
-                <strong>Email:</strong> {selectedIntern.email || "-"}
-              </p>
-              <p style={{ margin: 0 }}>
-                <strong>SĐT:</strong> {selectedIntern.phone || "-"}
-              </p>
-              <p style={{ margin: 0 }}>
-                <strong>Trường / Ngành:</strong> {selectedIntern.school || "-"} /{" "}
-                {selectedIntern.major || "-"}
-              </p>
-              <p style={{ margin: 0 }}>
-                <strong>Chương trình:</strong> {selectedIntern.programName || "-"}
-              </p>
-              <p style={{ margin: 0 }}>
-                <strong>Team / Mentor:</strong> {selectedIntern.teamName || "-"} /{" "}
-                {selectedIntern.mentorName || "-"}
-              </p>
-            </div>
-
-            <div
-              style={{
-                borderTop: "1px solid #f3f4f6",
-                paddingTop: 12,
-              }}
-            >
-              <h4
-                style={{
-                  margin: 0,
-                  fontSize: 16,
-                  fontWeight: 600,
-                  color: "#374151",
-                  marginBottom: 8,
-                }}
-              >
-                Thống kê đánh giá
-              </h4>
-              <div
-                style={{
-                  display: "grid",
-                  gridTemplateColumns: "repeat(3, minmax(0, 1fr))",
-                  gap: 8,
-                }}
-              >
-                <p style={{ margin: 0 }}>
-                  <strong>Số lần đánh giá:</strong>{" "}
-                  {selectedIntern.evaluationCount ?? 0}
-                </p>
-                <p style={{ margin: 0 }}>
-                  <strong>Điểm kỹ thuật:</strong>{" "}
-                  {selectedIntern.avgTechnical != null
-                    ? selectedIntern.avgTechnical.toFixed(2)
-                    : "-"}
-                </p>
-                <p style={{ margin: 0 }}>
-                  <strong>Giao tiếp:</strong>{" "}
-                  {selectedIntern.avgCommunication != null
-                    ? selectedIntern.avgCommunication.toFixed(2)
-                    : "-"}
-                </p>
-                <p style={{ margin: 0 }}>
-                  <strong>Kỷ luật:</strong>{" "}
-                  {selectedIntern.avgDiscipline != null
-                    ? selectedIntern.avgDiscipline.toFixed(2)
-                    : "-"}
-                </p>
-                <p style={{ margin: 0 }}>
-                  <strong>Thái độ:</strong>{" "}
-                  {selectedIntern.avgAttitude != null
-                    ? selectedIntern.avgAttitude.toFixed(2)
-                    : "-"}
-                </p>
-                <p style={{ margin: 0 }}>
-                  <strong>Điểm cuối kỳ:</strong>{" "}
-                  {selectedIntern.finalScore != null
-                    ? selectedIntern.finalScore.toFixed(2)
-                    : "-"}
-                </p>
-              </div>
-            </div>
-
-            <div
-              style={{
-                display: "flex",
-                flexDirection: "column",
-                gap: 12,
-              }}
-            >
-              <div>
-                <h4
-                  style={{
-                    margin: 0,
-                    fontSize: 15,
-                    fontWeight: 600,
-                    color: "#374151",
-                    marginBottom: 6,
-                  }}
-                >
-                  Ghi chú gần nhất
-                </h4>
-                <div
-                  style={{
-                    padding: "10px 12px",
-                    border: "1px solid #e5e7eb",
-                    borderRadius: 8,
-                    backgroundColor: "#f9fafb",
-                    minHeight: 40,
-                    fontSize: 14,
-                  }}
-                >
-                  {selectedIntern.latestNote || "Không có ghi chú"}
-                </div>
-              </div>
-
-              <div>
-                <h4
-                  style={{
-                    margin: 0,
-                    fontSize: 15,
-                    fontWeight: 600,
-                    color: "#374151",
-                    marginBottom: 6,
-                  }}
-                >
-                  Tất cả ghi chú
-                </h4>
-                <div
-                  style={{
-                    padding: "10px 12px",
-                    border: "1px solid #e5e7eb",
-                    borderRadius: 8,
-                    backgroundColor: "#f9fafb",
-                    minHeight: 60,
-                    whiteSpace: "pre-line",
-                    fontSize: 14,
-                  }}
-                >
-                  {selectedIntern.allNotes || "Không có ghi chú"}
-                </div>
-              </div>
-            </div>
-          </div>
-        </Modal>
-      )}
+      <InternDetailModal
+        selectedIntern={selectedIntern}
+        onClose={() => setSelectedIntern(null)}
+      />
     </div>
   );
 };
